requirements.txt
<<<<<<< HEAD
Flask
pytest
=======
aniso8601==8.1.0
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.1
Flask-RESTful==0.3.8
idna==2.10
inflection==0.5.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-spec-validator==0.2.9
psycopg2==2.8.6
pylint==2.6.0
pyrsistent==0.17.3
python-dotenv==0.19.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
swagger-ui-bundle==0.0.8
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
>>>>>>> origin/main
